Backend/requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-decouple>=3.8
Pillow>=10.0
gunicorn>=23.0
drf-spectacular>=0.27
